backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-filter>=23.5
django-cors-headers>=4.3
django-environ>=0.11
django-celery-beat>=2.5
django-celery-results>=2.5
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
beautifulsoup4>=4.12
requests>=2.31
lxml>=5.1
selenium>=4.15
